补充同步键位

+++ DES_GOBSTG/DES_GOBSTG/Header/Export.h
@@ -87,7 +87,7 @@ struct partInfo
 };
 
 struct replayFrame{
-	BYTE input;
+	WORD input;
 	BYTE bias;
 };
